requirements.txt
Flask
PyPDF2
python-docx
faiss-cpu
openai
werkzeug
gunicorn
flask-session
